modularise the code
