+++ index.html
@@ -41,6 +41,6 @@
       </form> 
   </div>
 
-  <script src="index.js"></script>
+  <script type = 'module' src="./js/main.js"></script>
 </body>
 </html>
