+++ js/main.js
@@ -1,33 +1,11 @@
 /* eslint-disable no-use-before-define */
-const myLibrary = [];
-const target = document.getElementById('demo');
-const targetbtn = document.getElementById('add-btn');
-const targetdisplay = document.getElementById('display');
-const alertText = document.getElementById('alertText');
+
+import {
+  myLibrary, target, targetbtn, targetdisplay, alertText, addBookToLibrary, customAlert,
+  displayForm,
+} from './library.js';
+
 let counter = 0;
-
-function Book(booktitle, author, totalPages, read) {
-  this.booktitle = booktitle;
-  this.author = author;
-  this.totalPages = totalPages;
-  this.read = read;
-}
-
-function saveLibrary() {
-  localStorage.setItem('myLibrary', JSON.stringify(myLibrary));
-}
-
-function addBookToLibrary(booktitle, author, totalPages, read = true) {
-  const book = new Book(booktitle, author, totalPages, read);
-  myLibrary.push(book);
-  saveLibrary();
-}
-
-function displayForm() {
-  targetbtn.style.display = 'none';
-  target.style.display = 'none';
-  targetdisplay.style.display = 'block';
-}
 
 function displayBooks() {
   let count = 0;
@@ -57,11 +35,8 @@ function displayBooks() {
 targetbtn.addEventListener('click', (e) => {
   e.preventDefault();
   displayForm();
+  console.log('okokok');
 });
-
-function customAlert() {
-  alertText.style.display = 'block';
-}
 
 document.querySelector('#bookForm').addEventListener('submit', (e) => {
   e.preventDefault();
